seperate combine bin stats from report
